+++ database/nostr-database-test-suite/src/lib.rs
@@ -40,7 +40,7 @@ macro_rules! database_unit_tests {
         }
 
         fn build_event(keys: &Keys, builder: EventBuilder) -> Event {
-            builder.sign_with_keys(keys).expect("Failed to build and sign event")
+            builder.finalize(keys).expect("Failed to build and sign event")
         }
 
         // Return the number of added events
@@ -82,7 +82,7 @@ macro_rules! database_unit_tests {
             builder: EventBuilder,
             keys: &Keys,
         ) -> (Event, SaveEventStatus) {
-            let event = builder.sign_with_keys(keys).expect("Failed to sign event");
+            let event = builder.finalize(keys).expect("Failed to sign event");
             let status = store.save_event(&event).await.expect("Failed to save event");
             (event, status)
         }
@@ -104,24 +104,24 @@ macro_rules! database_unit_tests {
             let helper = Keys::generate();
 
             let event1 = EventBuilder::text_note("Hi 1")
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
             let event2 = EventBuilder::text_note("Hi 2")
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
             let replaceable = EventBuilder::contact_list([
                 Contact::new(Keys::generate().public_key),
                 Contact::new(Keys::generate().public_key),
             ])
-            .sign_with_keys(&to_vanish)
+            .finalize(&to_vanish)
             .unwrap();
             let addresable = EventBuilder::long_form_text_note("LONG")
                 .tag(Tag::identifier("lorem-ipsum".to_string()))
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
             let dummy_gift_wrap = EventBuilder::new(Kind::GiftWrap, ":)")
                 .tag(Tag::public_key(to_vanish.public_key))
-                .sign_with_keys(&helper)
+                .finalize(&helper)
                 .unwrap();
 
             store.save_event(&event1).await.unwrap();
@@ -279,7 +279,7 @@ macro_rules! database_unit_tests {
             let metadata1 = Metadata::new().name("First");
             let event1 = EventBuilder::metadata(&metadata1)
                 .custom_created_at(Timestamp::from_secs(1000))
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
 
             store.save_event(&event1).await.expect("Failed to save event");
@@ -288,7 +288,7 @@ macro_rules! database_unit_tests {
             let metadata2 = Metadata::new().name("Second");
             let event2 = EventBuilder::metadata(&metadata2)
                 .custom_created_at(Timestamp::from_secs(2000))
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
 
             store.save_event(&event2).await.expect("Failed to save event");
@@ -313,7 +313,7 @@ macro_rules! database_unit_tests {
             let event1 = EventBuilder::new(Kind::from(32121), "Content 1")
                 .tag(Tag::identifier("test-id"))
                 .custom_created_at(Timestamp::from_secs(1000))
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
 
             store.save_event(&event1).await.expect("Failed to save event");
@@ -322,7 +322,7 @@ macro_rules! database_unit_tests {
             let event2 = EventBuilder::new(Kind::from(32121), "Content 2")
                 .tag(Tag::identifier("test-id"))
                 .custom_created_at(Timestamp::from_secs(2000))
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
 
             store.save_event(&event2).await.expect("Failed to save event");
@@ -347,10 +347,10 @@ macro_rules! database_unit_tests {
 
             // Create events to delete
             let event1 = EventBuilder::text_note("To be deleted 1")
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
             let event2 = EventBuilder::text_note("To be deleted 2")
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
 
             store.save_event(&event1).await.expect("Failed to save event");
@@ -359,7 +359,7 @@ macro_rules! database_unit_tests {
             // Create deletion event
             let deletion =
                 EventBuilder::delete(EventDeletionRequest::new().id(event1.id).id(event2.id))
-                    .sign_with_keys(&keys)
+                    .finalize(&keys)
                     .expect("Failed to sign");
 
             store.save_event(&deletion)
@@ -392,7 +392,7 @@ macro_rules! database_unit_tests {
 
             // Create events to delete
             let event = EventBuilder::new(Kind::Custom(11_111), "To be deleted 1")
-                .sign_with_keys(&keys)
+                .finalize(&keys)
                 .expect("Failed to sign");
 
             let status = store.save_event(&event).await.expect("Failed to save event");
@@ -404,7 +404,7 @@ macro_rules! database_unit_tests {
                 .coordinate(event.coordinate().unwrap().into_owned());
             let deletion =
                 EventBuilder::delete(req)
-                    .sign_with_keys(&keys)
+                    .finalize(&keys)
                     .expect("Failed to sign");
 
             store.save_event(&deletion)
@@ -945,7 +945,7 @@ macro_rules! database_unit_tests {
             // Request to vanish
             let request_to_vanish = EventBuilder::request_vanish(VanishTarget::AllRelays)
                 .unwrap()
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
             store.save_event(&request_to_vanish).await.unwrap();
 
@@ -970,7 +970,7 @@ macro_rules! database_unit_tests {
             );
 
             let new_event = EventBuilder::text_note("It was a mistake, please accept my event")
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
 
             // Try adding new event, should get rejected
@@ -996,7 +996,7 @@ macro_rules! database_unit_tests {
             // Request to vanish
             let request_to_vanish = EventBuilder::request_vanish(VanishTarget::relay(url))
                 .unwrap()
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
             store.save_event(&request_to_vanish).await.unwrap();
 
@@ -1021,7 +1021,7 @@ macro_rules! database_unit_tests {
             );
 
             let new_event = EventBuilder::text_note("It was a mistake, please accept my event")
-                .sign_with_keys(&to_vanish)
+                .finalize(&to_vanish)
                 .unwrap();
 
             // Try adding new event, should get rejected