+++ sdk/src/relay/inner.rs
@@ -1371,7 +1371,7 @@ impl InnerRelay {
 
         // Construct event
         let event: Event = EventBuilder::auth(challenge, self.url.clone())
-            .sign(signer)
+            .finalize_async(signer)
             .await?;
 
         // Subscribe to notifications